requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.6
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.13.2
esptool==2.6
idna==2.7
isort==4.3.20
jsonschema==2.6.0
lazy-object-proxy==1.4.1
Mastodon.py==1.4.6
mccabe==0.6.1
numpy==1.17.2
opencv-python==4.1.1.26
paho-mqtt==1.4.0
paramiko==2.5.0
Pillow==6.1.0
pyaes==1.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
PyNaCl==1.3.0
pyperclip==1.7.0
pyserial==3.4
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.2
PyYAML==3.13
requests==2.20.1
Roomba980-Python==1.2.3
six==1.12.0
texttable==0.9.1
typed-ast==1.4.0
urllib3==1.24.3
websocket-client==0.56.0
wrapt==1.11.1
youtube-dl==2019.9.1
